Move SpeakSettings classes to database module
(except save & load, which will stay as extensions)
