+++ app/src/main/java/net/bible/service/sword/SwordContentFacade.kt
@@ -22,8 +22,8 @@ import net.bible.android.BibleApplication.Companion.application
 import net.bible.android.activity.R
 import net.bible.android.control.ApplicationScope
 import net.bible.android.control.bookmark.BookmarkStyle
-import net.bible.android.control.speak.SpeakSettings
 import net.bible.android.control.versification.VersificationConverter
+import net.bible.android.database.bookmarks.SpeakSettings
 import net.bible.android.database.WorkspaceEntities.TextDisplaySettings
 import net.bible.service.common.CommonUtils.getResourceInteger
 import net.bible.service.common.CommonUtils.sharedPreferences
